fix width to unsigned

+++ src/metaSMT/backend/CVC4.hpp
@@ -213,7 +213,7 @@ namespace metaSMT {
       }
 
       result_type operator()( bvtags::zero_extend_tag const &
-        , uint64_t width
+        , unsigned width
         , result_type e)
       {
         ::CVC4::BitVectorZeroExtend bvOp (width);
@@ -222,7 +222,7 @@ namespace metaSMT {
       }
 
       result_type operator()( bvtags::sign_extend_tag const &
-        , uint64_t width
+        , unsigned width
         , result_type e)
       {
         ::CVC4::BitVectorSignExtend bvOp (width);
